requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.45
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.0
bcrypt==4.1.1
WTForms==3.1.1
email-validator==2.1.0
python-dateutil==2.8.2
requests==2.31.0
Werkzeug==3.0.1
